requirements.txt
requests==2.31.0
python-telegram-bot==21.4
Pillow==10.4.0
google-generativeai==0.8.3
asyncio